requirements.txt
fastapi
Pillow
pydantic
tensorflow
numpy
uvicorn
gunicorn==19.9.0
aiofiles==0.6.0
python-multipart
uvloop
httptools